Start implementing Programme Section on Home Page
Began the implementation of the Programme Section for Day 2. This includes setting up the structure for a keynote address, sessions, and breaks, along with a special event. Not all sections are completed, and further adjustments are needed to ensure the timeline marker dot only displays when relevant fields within each list item are not empty.