Improve card interactivity, add links and fix scrolling with open overlay menu
Added a clickable-parent class to anchor tags within cards, using a pseudo-element to expand the clickable area of the cards and a link to redirect users to a detailed single speaker page to see the abstract. Hid an image on sm and md screens within the Mission Section and prevented the page from scrolling when the overlay Menu is active.